Initial checkin for the MiOS Bridge Binding.
This binding exposes read, and read-command, access to Devices controlled by a MiOS Home Automation controller, such as those seen at http://getvera.com.

It exposes the ability to do the following things in the MiOS HA Controller

Devices - Read State Variables & Device Attributes, and invoke (single parameter) UPnP Commands to control the Device.
Scenes - Read the current execution state of a Scene, and invoke those Scenes within the remote HA Controller
System - Read System-level Attributes.
It uses the remote control interfaces (aka "UI Simple" JSON Calls, and HTTP Long-polling) of the MiOS HA Controller to keep the bound openHAB Items in sync with their counterparts in the MiOS HA Controller.

The "guts" is outlined in the root-level README.md file, along with the JavaDoc that's in the code (although, the README.md is better).

The binding uses the openHAB Transformation Service extensively to "map" the Data & Commands between the two systems. I provide a set of example MAP transform files in the examples/transform directory of the Binding, but these can readily be augmented without needing to tweak the code.

Original code was used from the XBMC Binding, and then heavily modified. Snippets included from the HTTP Binding for the various datatype mapping functions.