Squeezebox Binding: Sync "remove" players incorrect
The "sync:XXX" binding command for multi-player playback sync & its
respective OFF command is incorrect. It's removing the source player and
not the target player.
As the LMS "sync -" CLI command only supports a single playerID then it
should be the target synced player and not the source one.

Also means that when re-adding another player into another player group
the OFF command will always remove the correct respective player.