Add security mode setting to XMPP action
we are still a bit off from having secure XMPP connections in openhab,
since I expect most users are unable to provide a SSLContext that will
validate their certificate. But this, and the default setting of
'disabled', allows them at least to create a connection while being
informed that there is no connection security like TLS is used.