openhab-designer: set default swt.browser to mozilla
because using webkit ends with a segfault in libsoup.

This seems to be a common problem with this eclipse version and should
be fixed in newer versions. But according to the replies on the according
bug report: https://bugs.eclipse.org/bugs/show_bug.cgi?id=404776 this should
be a workaround that works for many people.

Signed-off-by: Manuel Traut <manut@mecka.net>