MiOS Binding: Only emit the Device Id IF at least one state-variable has been changed
... on the MiOS Device.

There are cases that MiOS sends out Device change records that contain no UPnP State Variable changes, they're effectively empty with the exception of the "device:nnnn/status" entry.  These are generating noise in the log files, so this change is to emit the Device Id field ONLY when there is at least one UPnP State Variable that's been changed.

Also, move the Device Id "device:nnnn/id" to the end of the list of attributes changed, in a given batch of changes for a Device.  This allows users to process Rules off the "batch of changes" as needed.  This can be handy if cross-Item validation Rules are present.

Reduce the # of log lines emitted during Debug mode, making the logs easier to read.

Use local Device and Scene caches for "status" attribute to avoid duplicate changes in openHAB.  MiOS has a bug where it'll repeatedly send the same "status" value for Devices and Scenes, even when they haven't changed.  This will result in many (many!) duplicate values being published into openHAB, and the resulting Rule executions and Log output.  This change addresses that problem by retaining local Device and Scene caches for the "status" attribute (only).  We cannot use openHAB/ItemRegistry here, since values are applied async, and we could get ourselves into an inconsistent state.  The reduction in Log output (events.log) alone is about 85% in my system, which greatly reduces the IO write contention on rPi and Odroid Flash filesystems.