Edit Aeon Multi-Sensor 6
Added parameter 7 but am awaiting feedback on other issues. I am
hesitant to remove what may appear to be superfluous parameters as
documentation does refer to some of those values. As for adding a
parameter to set 101 - 103 to default there are no values given in the
documentation for doing that. I will await feedback on affecting those
changes as well.