[silvercrest] 1952 add easy home to silver crest (#1954)
* Enrich SilverCrest Binding with EasyHome functionality

Added functionality to handle EasyHome equipment with this binding as
well. Main change is the addition of a vendor field and property and an
Enum to represent different versions of the similar hardware. The main
change of packet composing/decomposing is in WifiSocketPacketConverter.
Included review comments by Martin van Wingerden

Also-by: Martin van Wingerden <martinvw@mtin.nl>
Signed-off-by: Christian Heimerl <Christian.Heimerl@gmail.com> (github: Ronson1909)