Fix powerview configuration to use String for id (#2521)
In 2.0, referencing a configuration defined with type 'text' as an int was allowed, so long as the value was initialized as an int originally. In 2.1, this no longer works.

Updating all references to the config to a String. Doing that instead of just changing the thing-type in order to preserve user's configuration. We don't really care about the type anyways - it's just a unique identifier for us.

Fixes #2520

Signed-off-by: Andy Lintner dev@beowulfe.com