Don't cache data points of reconfigurable channels. (#2932)
Channels that allow changing their type (e.g. the KEY channels in
HM-MOD-EM-8) have different data point sets depending on the selected
function. If multiple of such devices are present and are configured to
have differing functions, caching the data points of those channels can
lead to us enumerating wrong data points, as the configuration of the
first device will be implied for all subsequent devices.

Closes #2930

Signed-off-by: Danny Baumann <dannybaumann@web.de>